Make aggregator an enum and retention an object
Aggregator used to be a string named carbon_aggregation, and retention a
list of (point, precision).
This API-only change makes it easier to do server-side aggregation.
It also removes the support for xfilesfactor which we don't use and
won't support with server-side aggregation at first (we still store it
though)