cassandra: change time_offset_ms to offset (which unit is stage.precision_ms)
time_offset_ms was limited to 46 days (which doesn't work well when
you store >1000 points with a resolution of one day in a single row).

To overcome this limitation, change offset to be related to stage.precision_ms
which allows us to store up to 4B points in a single row (which we probably
won't do, and we will use the high bits for other purposes later).