cassanra: separate stage0 and aggregated metrics
The first stage of a retention policy is called 'stage0'.
Other stages are 'aggregated' which means that they also keep
track of the number of points aggregated in the value because
carbon doesn't read last aggregate we also need to track
the writers to avoid data loss on restart.

'stage0' and 'aggregated' also have different kind of traffic
and can benefit from custom settings (different memtable_flush_period
values, etc..)