cassandra / graphite: better caching for find() queries
Instead of caching only globs we directly cache CQL queries
in a shared cache. This is interesting because N globs can map
to the same CQL query (we don't have support for suffix search).

On top of that we also cache exceptions to make sure we cache long-running
and broken queries.