cassandra: ignore timeout in background jobs
We can get timeouts from time to time, and that's fine. We should
not stop the whole job because we got a single timeout. This patch
is trying to make the background operations more robust. Simple errors
will simply be ignored and the code will continue to the next range.

- Add a retry policy that always retries (3 times by default).
  This needs https://github.com/datastax/python-driver/pull/861 to work.
- Change the consistency to ONE when querying indexes (because indexes are
  expensive to query).
- Ignore ReadTimeout and HostUnavailable.