drivers: make sure touch_metric() also changes the metric
Currently it changes the database but not the object used
as a parameter. Since both are not synced, the next call
will re-trigger a database update even if there was no need
for it.

In elasticsearch also fetch the document only when we really
need, else we end up doing one synchronous query each time
we write a point.