Fix metadata read parsing in Cassandra driver.
This patch fix an bug in the parsing of metric metadata when gettting it
directly from cassandra.
Updated_on is return as timestamp and was parsed as int; but we expect
a datetime.datetime object.
This also add a test trying to write datapoint from an existing metric
and without using the cache. This test would have catch this issue.

Example error log:
```07/08/2018 17:53:21 :: [console] Error writing to carbon.agents.tchataigner-laptop-a.process_virtual_memory_bytes:
'int' object has no attribute 'timetuple'```

Signed-off-by: Thibault Chataigner <t.chataigner@criteo.com>