Cleaning improvement & bugfixes (#557)
* Fixing Travis build.

* Enhance cleaner's progressbar & metrics.

Fixing progressbar so output is better (with current token!)
Allow exporting metrics in a single file, so they can be exported
using a node_exporter.

* Do not crash if updated_on is missing when retrieving metadata.

* Drop cache on not found metric error.

When cleaning cassandra backend, some inconsistency might appear,
leaving the cache in a state that a metric is cached with None
content. Dropping the cache will force the metric to be recreated
& repopulated correctly.

* Cleaner: Widen the scope of catch'ed exceptions.

On some case, it happened that we lost the cleaner because of too much
node timeout & retries. Ignoring those are warmless.