Dump all metrics names on disk when cleaning.
When cleaning metrics, we're sequentialy listing all metrics in the
unfiltered way to do so. Therefore, adding a way to extract all those
for debug/catalog reasons.
This is not enabled by default, and is triggered only when method
is 'unfiltered' and the env. var CLEAN_DUMP_METRICS is set.

While I'm at it, disabling python 2 builds & checks.